Issue 69: remove Japanese from new db if user doesn't have MeCab.
The demo db used as the DB baseline always has Japanese language
and a JP sample story, so it would try to parse it, throwing an
error.  It's too messy to load the JP story at runtime, due to
various dependency injections, so we just remove the unusable
data instead.